proof_of_concept/requirements.txt
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.28
alembic==1.13.1
pyyaml==6.0.1
tkinter==0.1.0
matplotlib==3.8.2
numpy==1.26.3
pandas==2.1.4
psutil==5.9.6
tqdm==4.66.1
pytest==7.4.3
pytest-asyncio==0.21.1
docker==6.1.3